+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/bootstrap/domain/BootstrapUseCase.kt
@@ -1,14 +1,14 @@
 package io.github.composegears.valkyrie.ui.screen.webimport.standard.bootstrap.domain
 
 import io.github.composegears.valkyrie.settings.InMemorySettings
+import io.github.composegears.valkyrie.settings.toPersistedString
+import io.github.composegears.valkyrie.settings.toStringList
 import io.github.composegears.valkyrie.ui.screen.webimport.common.model.FontByteArray
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.bootstrap.data.BootstrapRepository
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.StandardIconProvider
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.SvgSizeCustomizer
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.inferCategoryFromTags
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.toDisplayName
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.IconStyle
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.SizeSettings
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.StandardIcon
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.StandardIconConfig
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.toStandardIconConfig
@@ -22,10 +22,19 @@ class BootstrapUseCase(
     override val stateKey: String = "bootstrap"
     override val fontAlias: String = "bootstrap-icons"
     override val persistentSize: Int = inMemorySettings.readState { bootstrapSize }
+    override val persistentLastCustomColor: String? = inMemorySettings.readState { bootstrapLastCustomColor }?.ifBlank { null }
+    override val persistentRecentColors: List<String> = inMemorySettings.readState { bootstrapRecentColors }.toStringList()
 
     override fun updatePersistentSize(value: Int) {
         inMemorySettings.update {
             bootstrapSize = value
+        }
+    }
+
+    override fun updatePersistentCustomColors(lastCustomColor: String?, recentColors: List<String>) {
+        inMemorySettings.update {
+            bootstrapLastCustomColor = lastCustomColor.orEmpty()
+            bootstrapRecentColors = recentColors.toPersistedString()
         }
     }
 
@@ -49,8 +58,5 @@ class BootstrapUseCase(
         return FontByteArray(repository.loadFontBytes())
     }
 
-    override suspend fun downloadSvg(icon: StandardIcon, settings: SizeSettings): String {
-        val rawSvg = repository.downloadSvg(icon.name)
-        return SvgSizeCustomizer.applySettings(rawSvg, settings)
-    }
+    override suspend fun loadSvg(icon: StandardIcon): String = repository.downloadSvg(icon.name)
 }