+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/standard/common/StandardImportScreenUI.kt
@@ -46,13 +46,13 @@ import io.github.composegears.valkyrie.ui.screen.webimport.common.ui.CategoryHea
 import io.github.composegears.valkyrie.ui.screen.webimport.common.ui.IconCard
 import io.github.composegears.valkyrie.ui.screen.webimport.common.ui.IconGrid
 import io.github.composegears.valkyrie.ui.screen.webimport.common.ui.IconLoadingPlaceholder
-import io.github.composegears.valkyrie.ui.screen.webimport.common.ui.IconSizeCustomization
 import io.github.composegears.valkyrie.ui.screen.webimport.common.ui.SidePanel
+import io.github.composegears.valkyrie.ui.screen.webimport.common.ui.SvgCustomizationPanel
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.domain.StandardIconProvider
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.IconStyle
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.InferredCategory
-import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.SizeSettings
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.StandardIcon
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.model.SvgImportSettings
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.common.ui.StandardTopActions
 import io.github.composegears.valkyrie.util.stringResource
 import kotlinx.coroutines.launch
@@ -91,6 +91,8 @@ internal fun StandardImportScreen(
         onSelectStyle = viewModel::selectStyle,
         onSearchQueryChange = viewModel::updateSearchQuery,
         onSettingsChange = viewModel::updateSettings,
+        onPickCustomColor = viewModel::selectCustomColor,
+        onResetCustomization = viewModel::resetCustomization,
         modifier = modifier,
     )
 }
@@ -106,7 +108,9 @@ private fun StandardImportScreenUI(
     onSelectCategory: (InferredCategory) -> Unit,
     onSelectStyle: (IconStyle) -> Unit,
     onSearchQueryChange: (String) -> Unit,
-    onSettingsChange: (SizeSettings) -> Unit,
+    onSettingsChange: (SvgImportSettings) -> Unit,
+    onPickCustomColor: (String) -> Unit,
+    onResetCustomization: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     Column(
@@ -155,6 +159,8 @@ private fun StandardImportScreenUI(
                         onSelectStyle = onSelectStyle,
                         onSearchQueryChange = onSearchQueryChange,
                         onSettingsChange = onSettingsChange,
+                        onPickCustomColor = onPickCustomColor,
+                        onResetCustomization = onResetCustomization,
                     )
                 }
             }
@@ -171,7 +177,9 @@ private fun IconsContent(
     onSelectCategory: (InferredCategory) -> Unit,
     onSelectStyle: (IconStyle) -> Unit,
     onSearchQueryChange: (String) -> Unit,
-    onSettingsChange: (SizeSettings) -> Unit,
+    onSettingsChange: (SvgImportSettings) -> Unit,
+    onPickCustomColor: (String) -> Unit,
+    onResetCustomization: () -> Unit,
 ) {
     val scope = rememberCoroutineScope()
 
@@ -284,11 +292,15 @@ private fun IconsContent(
             isOpen = isSidePanelOpen,
             onClose = { isSidePanelOpen = false },
             content = {
-                IconSizeCustomization(
+                SvgCustomizationPanel(
                     settings = state.settings,
+                    recentColors = state.recentColors,
+                    lastCustomColor = state.lastCustomColor,
+                    capabilities = state.customizationCapabilities,
                     onSettingsChange = onSettingsChange,
+                    onPickCustomColor = onPickCustomColor,
+                    onResetCustomization = onResetCustomization,
                     onClose = { isSidePanelOpen = false },
-                    sizeLabel = stringResource("web.import.font.customize.size"),
                 )
             },
         )